Moved CDNCache from cdn_cache.py to tasks/common.py
CDNCache provides flush_cdn functionality which is common
across multiple tasks. The method is more suited to be used
along with repo related information available in the task.
Hence, looks more suitable to be placed under tasks.
