+++ pubtools/_pulp/tasks/clear_repo.py
@@ -14,12 +14,12 @@ from pubtools.pulplib import (
 )
 
 from pubtools._pulp.task import PulpTask
-from pubtools._pulp.cdn_cache import CDNCache
 from pubtools._pulp.services import (
     CollectorService,
     UdCacheClientService,
     PulpClientService,
 )
+from pubtools._pulp.tasks.common import CDNCache
 
 step = PulpTask.step
